Changed to use long value to generate seed

+++ snake3d/src/main/java/com/sakariaslilja/App.java
@@ -18,7 +18,7 @@ import com.sakariaslilja.services.GameEngine;
 public class App extends Application {
 
     private static Scene scene;
-    private static GameEngine engine = new GameEngine((int) System.nanoTime(), 3, 3, 3, new DoubleVector3D(0.5, 0.5, 0.5), 0, 0, 0);
+    private static GameEngine engine = new GameEngine(System.nanoTime(), 3, 3, 3, new DoubleVector3D(0.5, 0.5, 0.5), 0, 0, 0);
 
     // Variables of the scene.
     private static final String primarySceneFXML = "mainpage";
